typo updated on syllabus
Hi Leslie! 
I learned how to branch repos this weekend and practiced a little bit before giving this a try. One of the subtitles was spelled incorrectly (Preaparing instead of Preparing) and instead of sending you a slack message or something similar, I thought I should audaciously give this a try and submit a pull request to update the syllabus.